rename current v3 to v4
