+++ crates/skytem_logs/src/mbed_motor_control/status/header.rs
@@ -4,7 +4,7 @@ use byteorder::{LittleEndian, ReadBytesExt};
 use serde::{Deserialize, Serialize};
 use v1::StatusLogHeaderV1;
 use v2::StatusLogHeaderV2;
-use v3::StatusLogHeaderV3;
+use v4::StatusLogHeaderV4;
 
 use crate::{
     mbed_motor_control::mbed_header::{
@@ -15,13 +15,13 @@ use crate::{
 
 mod v1;
 mod v2;
-mod v3;
+mod v4;
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub enum StatusLogHeader {
     V1(StatusLogHeaderV1),
     V2(StatusLogHeaderV2),
-    V3(StatusLogHeaderV3),
+    V3(StatusLogHeaderV4),
 }
 
 impl fmt::Display for StatusLogHeader {
@@ -75,8 +75,8 @@ impl StatusLogHeader {
                 total_bytes_read += bytes_read;
                 Self::V2(header)
             }
-            3 => {
-                let (header, bytes_read) = StatusLogHeaderV3::from_reader_with_uniq_descr_version(
+            4 => {
+                let (header, bytes_read) = StatusLogHeaderV4::from_reader_with_uniq_descr_version(
                     reader,
                     unique_description,
                     version,
